# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Todo API",
    "description": "REST API для управления задачами (Todos) и пользователями (Users).\nПоддерживает полный CRUD для обеих сущностей.",
    "version": "1.0.0",
    "termsOfService": "https://opensource.org/licenses/MIT",
    "contact": {
      "name": "Netology AI-Assist Team",
      "email": "support@netology.ru"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "tags": [
    {
      "name": "users",
      "description": "Управление пользователями"
    },
    {
      "name": "todos",
      "description": "Управление задачами"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "List all users",
        "description": "Returns a list of all registered users wrapped in SuccessResponse",
        "operationId": "ListUsers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "List of users",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create a new user",
        "description": "Creates a new user with the provided email and name",
        "operationId": "CreateUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "User data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Get a user by ID",
        "description": "Returns a single user by their ID wrapped in SuccessResponse",
        "operationId": "GetUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update an existing user",
        "description": "Updates a user by their ID. Both email and name are optional.",
        "operationId": "UpdateUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "request",
            "description": "Updated user data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete a user",
        "description": "Deletes a user by their ID",
        "operationId": "DeleteUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": ["todos"],
        "summary": "List all tasks",
        "description": "Returns a list of all tasks wrapped in SuccessResponse",
        "operationId": "ListTodos",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "List of tasks",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Todo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["todos"],
        "summary": "Create a new task",
        "description": "Creates a new task with the provided data",
        "operationId": "CreateTodo",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "Task data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTodoRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Task created",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/definitions/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "tags": ["todos"],
        "summary": "Get a task by ID",
        "description": "Returns a single task by its ID wrapped in SuccessResponse",
        "operationId": "GetTodo",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/definitions/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Task not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": ["todos"],
        "summary": "Update an existing task",
        "description": "Updates a task by its ID. All fields are optional.",
        "operationId": "UpdateTodo",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "request",
            "description": "Updated task data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateTodoRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task updated",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/definitions/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Task not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["todos"],
        "summary": "Delete a task",
        "description": "Deletes a task by its ID",
        "operationId": "DeleteTodo",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Task not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "description": "Represents a single task with its metadata",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "Buy groceries"
        },
        "description": {
          "type": "string",
          "example": "Milk, bread, eggs, cheese"
        },
        "done": {
          "type": "boolean",
          "example": false
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CreateTodoRequest": {
      "type": "object",
      "description": "Request body for creating a new task",
      "required": ["user_id", "title"],
      "properties": {
        "user_id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "title": {
          "type": "string",
          "maxLength": 200,
          "example": "Buy groceries"
        },
        "description": {
          "type": "string",
          "maxLength": 1000,
          "example": "Milk, bread, eggs, cheese"
        },
        "done": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "UpdateTodoRequest": {
      "type": "object",
      "description": "Request body for updating an existing task (all fields optional)",
      "properties": {
        "title": {
          "type": "string",
          "maxLength": 200,
          "example": "Buy groceries and snacks"
        },
        "description": {
          "type": "string",
          "maxLength": 1000,
          "example": "Updated description"
        },
        "done": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "User": {
      "type": "object",
      "description": "Represents a user account",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "email": {
          "type": "string",
          "format": "email",
          "example": "user@example.com"
        },
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "description": "Request body for creating a new user",
      "required": ["email", "name"],
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "user@example.com"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "example": "John Doe"
        }
      }
    },
    "UpdateUserRequest": {
      "type": "object",
      "description": "Request body for updating an existing user (all fields optional)",
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "newemail@example.com"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "example": "Jane Doe"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "description": "Standard error response",
      "properties": {
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}